Added new hack PacketMine as a result of development efforts on FastBreak
Added PacketMine hack as a result of development efforts on FastBreak. It makes so that the player cannot mine blocks and instead mining is done by the hack by sending packets. The user has a visual representation of the enqueued blocks that he choose to mine and the blocks that are currently being mined by the hack.